First version of the preview before commit feature.
The metadata edition views have been refactored to move
as much logic as possible to the forms. As a result the
views code has been greatly simplified and is much more
readable. The forms do not have too much complexity since
we can use inheritance to share common code.

In the client side there is a new jQuery plugin called
mesh (Metadata Editor Shell) that validates the changes
and show them in a modal dialog before asking the user
for a commit message.